task 0 annotated v3

+++ python_async_function/0-basic_async_syntax.py
@@ -7,7 +7,7 @@ import asyncio
 import random
 
 
-async def wait_random(max_delay: float = 10) -> float:
+async def wait_random(max_delay: float = 10) -> int:
     """Generate a random delay using random.uniform(0, max_delay)"""
     delay: float = random.uniform(0, max_delay)
     await asyncio.sleep(delay)
